Added SampleDataset
Added a sample dataset which gets added/updated when you log in with a
new account. It added the account type (Google or Facebook currently)
along with the access token. This is just for demonstration purposes
and should not be included in a production app as storing the access
tokens in the dataset is probably a bad idea (it’s visible in the
Amazon Cognito dashboard).